% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAMax.R
\docType{methods}
\name{addStats,PCAMax-method}
\alias{addStats,PCAMax-method}
\alias{addStats}
\title{add PCA-based summary stats on all TFs in the model}
\usage{
\S4method{addStats}{PCAMax}(obj)
}
\arguments{
\item{obj}{An object of the class PCAMax}
}
\value{
the original model with extra columns: pcaMax, cov, PC1, PC2,
}
\description{
add PCA-based summary stats on all TFs in the model
}
