% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAMax.R
\docType{methods}
\name{addStatsSimple,PCAMax-method}
\alias{addStatsSimple,PCAMax-method}
\alias{addStatsSimple}
\title{add PCA-based summary stats on all TFs in the model}
\usage{
\S4method{addStatsSimple}{PCAMax}(obj, varianceToInclude = 0.75,
  scalePCA = FALSE, excludeLasso = TRUE, quiet = TRUE)
}
\arguments{
\item{obj}{An object of the class PCAMax}

\item{varianceToInclude}{numeric variance to include in the PCA}

\item{scalePCA}{logical}

\item{excludeLasso}{logical excluding lasso avoids dropping TFs due to shrinkage}

\item{quiet}{logical}
}
\value{
the original model with extra columns: pcaMax, cov, PC1, PC2,
}
\description{
add PCA-based summary stats on all TFs in the model
}
