% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpikeSolver.R
\docType{class}
\name{BayesSpikeSolver-class}
\alias{BayesSpikeSolver-class}
\alias{.BayesSpikeSolver}
\title{An S4 class to represent a Bayes Spike solver}
\description{
An S4 class to represent a Bayes Spike solver
}
