% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicorSolver.R
\name{BicorSolver}
\alias{BicorSolver}
\title{Create a Solver class object using  Bicor correlation coefficients as the solver}
\usage{
BicorSolver(
  mtx.assay = matrix(),
  targetGene,
  candidateRegulators,
  quiet = TRUE
)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{targetGene}{A designated target gene that should be part of the mtx.assay data}

\item{candidateRegulators}{The designated set of transcription factors that could be associated
with the target gene}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with Bicor correlation coefficients as the solver
}
\description{
Create a Solver class object using  Bicor correlation coefficients as the solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
bicor.solver <- BicorSolver(mtx.sub, target.gene, tfs)
}
\seealso{
\code{\link{solve.Bicor}}, \code{\link{getAssayData}}

Other Solver class objects: 
\code{\link{EnsembleSolver}},
\code{\link{HumanDHSFilter-class}},
\code{\link{LassoPVSolver}},
\code{\link{LassoSolver}},
\code{\link{PearsonSolver}},
\code{\link{RandomForestSolver}},
\code{\link{RidgeSolver}},
\code{\link{Solver-class}},
\code{\link{SpearmanSolver}},
\code{\link{XGBoostSolver}}
}
\concept{Solver class objects}
