% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateFilter.R
\docType{class}
\name{CandidateFilter-class}
\alias{CandidateFilter-class}
\alias{.CandidateFilter}
\alias{CandidateFilter}
\alias{CandidateFilter}
\title{CandidateFilter}
\usage{
CandidateFilter(quiet = TRUE)
}
\arguments{
\item{quiet}{A logical denoting whether or not the CandidateFilter object should print output}
}
\value{
An object of the Candidate filter class
}
\description{
A CandidateFilter is an S4 class to represent a gene candidate filter. These filters can employ
a variety of methods to reduce the number of transcription factors used as predictors for solving
a Solver object.
}
\section{Slots}{

\describe{
\item{\code{quiet}}{A logical denoting whether or not the CandidateFilter object should print output}
}}

\examples{
# Create an empty candidate filter
candidate.filter <- CandidateFilter(quiet=TRUE)
}
\seealso{
\code{\link{getCandidates}}
}
