% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{closeDatabaseConnections,FootprintFinder-method}
\alias{closeDatabaseConnections,FootprintFinder-method}
\alias{closeDatabaseConnections}
\title{Close a Footprint Database Connection}
\usage{
\S4method{closeDatabaseConnections}{FootprintFinder}(obj)
}
\arguments{
\item{obj}{An object of class FootprintFinder}
}
\value{
Closes the specified database connection
}
\description{
This method takes a FootprintFinder object and closes connections to the footprint databases
if they are currently open.
}
\seealso{
Other FootprintFinder methods: \code{\link{FootprintFinder-class}},
  \code{\link{getChromLoc,FootprintFinder-method}},
  \code{\link{getFootprintsForGene,FootprintFinder-method}},
  \code{\link{getFootprintsInRegion,FootprintFinder-method}},
  \code{\link{getGenePromoterRegion,FootprintFinder-method}},
  \code{\link{getGtfGeneBioTypes,FootprintFinder-method}},
  \code{\link{getGtfMoleculeTypes,FootprintFinder-method}},
  \code{\link{getPromoterRegionsAllGenes,FootprintFinder-method}},
  \code{\link{mapMotifsToTFsMergeIntoTable,FootprintFinder-method}}
}
