% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trena.R
\name{createGeneModel}
\alias{createGeneModel}
\title{Create a model for a target gene using a Trena object}
\usage{
createGeneModel(obj, targetGene, solverNames, tbl.regulatoryRegions, mtx)
}
\arguments{
\item{obj}{An object of class Trena}

\item{targetGene}{The name of a target gene to use for building a model}

\item{solverNames}{A character vector containing the solver names to be used for building the model}

\item{tbl.regulatoryRegions}{A data frame of regulatory regions, typically generated by using a filter}

\item{mtx}{An assay matrix of expression data}
}
\value{
A data frame containing the gene model
}
\description{
Create a model for a target gene using a Trena object
}
\examples{
# Create a Trena object for human and make a gene model for "MEF2C" using a footprint filter
trena <- Trena("hg38")
chromosome <- "chr5"
mef2c.tss <- 88904257
loc.start <- mef2c.tss - 1000
loc.end   <- mef2c.tss + 1000
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))

regions <- getRegulatoryChromosomalRegions(trena, chromosome, mef2c.tss-1000, mef2c.tss+1000,
sources, "MEF2C", mef2c.tss)
regions.tf <- expandRegulatoryRegionsTableByTF(trena, regions)
model.mef2c <- createGeneModel(trena, "MEF2C", c("lasso","ridge","randomforest"), regions.tf, mtx.sub) 
}
