% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trena.R
\docType{methods}
\name{createGeneModelFromTfList,Trena-method}
\alias{createGeneModelFromTfList,Trena-method}
\alias{createGeneModelFromTfList}
\title{Create a model for a target gene using a Trena object}
\usage{
\S4method{createGeneModelFromTfList}{Trena}(obj, targetGene, solverNames,
  tfList, mtx)
}
\arguments{
\item{obj}{An object of class Trena}

\item{targetGene}{The name of a target gene to use for building a model}

\item{solverNames}{A character vector containing the solver names to be used for building the model}

\item{tfList}{A character list, often the gene symbols for known transcription factors}

\item{mtx}{An assay matrix of expression data}
}
\value{
A data frame containing the gene model
}
\description{
Create a model for a target gene using a Trena object
}
\examples{
if(interactive()){  # takes too long for the bioconductor build
   # Create a Trena object for human and make a gene model for "MEF2C" using a footprint filter
   trena <- Trena("hg38")
   chromosome <- "chr5"
   mef2c.tss <- 88904257
   loc.start <- mef2c.tss - 1000
   loc.end   <- mef2c.tss + 1000

   database.filename <- system.file(package="trena", "extdata", "mef2c.neigborhood.hg38.footprints.db")
   database.uri <- sprintf("sqlite://\%s", database.filename)
   sources <- c(database.uri)
   load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))

   model.mef2c <- createGeneModelFromTfList(trena, "MEF2C", c("lasso","ridge","randomforest"),
                                            tfList=c())
   } # if interactive
}
