% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedFunctions.R
\name{elasticNetSolver}
\alias{elasticNetSolver}
\title{Run the Elastic Net Solvers}
\usage{
elasticNetSolver(obj, target.gene, tfs, tf.weights, alpha, lambda, keep.metrics)
}
\arguments{
\item{obj}{An object of class Solver}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}

\item{alpha}{The LASSO/Ridge tuning parameter}

\item{lambda}{The penalty tuning parameter for elastic net}

\item{keep.metrics}{A binary variable indicating whether or not to keep metrics}
}
\value{
A data frame containing the coefficients relating the target gene to each transcription factor, plus other fit parameters
}
\description{
Given a TReNA object with either LASSO or Ridge Regression as the solver, use the \code{\link{glmnet}} function to estimate coefficients for each transcription factor as a predictor of the target gene's expression level.
}
\seealso{
\code{\link{glmnet}}
}
