% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trena.R
\name{expandRegulatoryRegionsTableByTF}
\alias{expandRegulatoryRegionsTableByTF}
\title{Add transcription factors to a table of regulatory regions for a Trena object}
\usage{
expandRegulatoryRegionsTableByTF(obj, tbl.reg)
}
\arguments{
\item{obj}{An object of class Trena}

\item{tbl.reg}{A data frame of regulatory regions, typically generated by using a filter}
}
\value{
A data frame containing the gene model
}
\description{
Add transcription factors to a table of regulatory regions for a Trena object
}
\examples{
# Create a Trena object for human, retrieve regions for "MEF2C" using a footprint filter,
# and then add the TF column
trena <- Trena("hg38")
chromosome <- "chr5"
mef2c.tss <- 88904257
loc.start <- mef2c.tss - 1000
loc.end   <- mef2c.tss + 1000

regions <- getRegulatoryChromosomalRegions(trena, chromosome, mef2c.tss-1000, mef2c.tss+1000,
sources, "MEF2C", mef2c.tss)
regions.tf <- expandRegulatoryRegionsTableByTF(trena, regions)
}
