% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureTable.R
\name{FeatureTable}
\alias{FeatureTable}
\title{FeatureTable}
\description{
A template for building documented, tested R6 classes
}
\examples{
  rt <- FeatureTable()

import R6
import data.table

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureTable-new}{\code{FeatureTable$new()}}
\item \href{#method-FeatureTable-getTable}{\code{FeatureTable$getTable()}}
\item \href{#method-FeatureTable-setFundamentalRegions}{\code{FeatureTable$setFundamentalRegions()}}
\item \href{#method-FeatureTable-addFeature}{\code{FeatureTable$addFeature()}}
\item \href{#method-FeatureTable-clone}{\code{FeatureTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$new(target.gene, reference.genome)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.gene}}{character, a recognized gene symbol}

\item{\code{reference.genome}}{character, a recognized genome shorthand code, e.g., "hg38"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance o FeatureTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-getTable"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-getTable}{}}}
\subsection{Method \code{getTable()}}{
accessor for the table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$getTable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-setFundamentalRegions"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-setFundamentalRegions}{}}}
\subsection{Method \code{setFundamentalRegions()}}{
establish the basic genomic regions, typically
  transcription factor binding sites
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$setFundamentalRegions(tbl.regions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl.regions}}{a data.frame with at least chrom, start end columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-addFeature"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-addFeature}{}}}
\subsection{Method \code{addFeature()}}{
annotate each of the intersecting fundamental regions with
  the value of this feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$addFeature(tbl.feature, feature.genome, feature.guide)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature.genome}}{character, a recognized genome shorthand code, e.g., "hg38"}

\item{\code{tbl.bed}}{a data.frame with chrom, start and end columns, and values
of interest described in the annotation guidel}

\item{\code{annotation.guide}}{a neme list identifying columns of interest in
tbl.bed, and the column names to use for them in the feature table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
