% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifMatcher.R
\docType{methods}
\name{findMatchesByChromosomalRegion,MotifMatcher-method}
\alias{findMatchesByChromosomalRegion,MotifMatcher-method}
\title{Find Motif Matches by Chromosomal Region}
\usage{
\S4method{findMatchesByChromosomalRegion}{MotifMatcher}(obj, tbl.regions,
  pwmMatchMinimumAsPercentage, variants = NA_character_)
}
\arguments{
\item{obj}{An object of class MotifMatcher}

\item{tbl.regions}{A data frame where each row contains a chromosomal region with the fields
"chrom", "start", and "end".}

\item{pwmMatchMinimumAsPercentage}{A percentage (0-100) used as a cutoff for what constitutes
a motif match}

\item{variants}{A character containing variants to use for the matching (default = NA_character_).
The variants should either have the same number of entries as rows in the \code{tbl.regions},
or they should not be supplied.}
}
\value{
A list containing a data frame of the motifs in the given regions and a character
vector of their associated transcription factors
}
\description{
Given a MotifMatcher object, a table of chromosomal regions, and a minimum match percentage,
pull out a list containing a data frame of the motifs in those regions and a character vector
of their associated transcription factors.
}
\examples{
# Perform a simple match in the rs13384219 neighborhood
motifMatcher <- MotifMatcher(name="rs13384219.neighborhood", genomeName="hg38", quiet=FALSE)
tbl.regions <- data.frame(chrom="chr2", start=57907313, end=57907333, stringsAsFactors=FALSE)
x <- findMatchesByChromosomalRegion(motifMatcher, tbl.regions, pwmMatchMinimumAsPercentage=92)

# Perform the same match, but now include a variant
x.mut <- findMatchesByChromosomalRegion(motifMatcher, tbl.regions,
pwmMatchMinimumAsPercentage=92, "rs13384219")
}
