% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneOntologyFilter.R
\name{GeneOntologyFilter}
\alias{GeneOntologyFilter}
\title{Create a CandidateFilter using Gene Ontology}
\usage{
GeneOntologyFilter(organismDatabase = org.Hs.eg.db::org.Hs.eg.db,
  GOTerm = "GO:0006351", quiet = TRUE)
}
\arguments{
\item{organismDatabase}{An organism-specific database of type 'OrgDb'}

\item{GOTerm}{A character matching an accepted gene ontology term. The default term corresponds
to "transcription, DNA-templated". (default="GO:0006351")}

\item{quiet}{A logical denoting whether or not the filter should print output}
}
\value{
A GeneOntologyFilter object
}
\description{
Create a CandidateFilter using Gene Ontology
}
\examples{
# Make a filter for "transcription, DNA-templated"
library(org.Hs.eg.db)
goFilter <- GeneOntologyFilter(org.Hs.eg.db, GOTerm="GO:0006351")
}
\seealso{
\code{\link{CandidateFilter}}
}
