% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\docType{methods}
\name{getAssayData,Solver-method}
\alias{getAssayData,Solver-method}
\alias{getAssayData}
\title{Retrieve the assay matrix of gene expression data from a Solver object}
\usage{
\S4method{getAssayData}{Solver}(obj)
}
\arguments{
\item{obj}{An object of class Solver}
}
\value{
The assay matrix of gene expression data associated with a Solver object
}
\description{
Retrieve the assay matrix of gene expression data from a Solver object
}
\examples{
# Create a Solver object using the included Alzheimer's data and retrieve the matrix
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "MEF2C"
candidateRegulators <- setdiff(rownames(mtx.sub), targetGene)
solver <- Solver(mtx.sub, targetGene, candidateRegulators)
mtx <- getAssayData(solver)
}
