% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAvailableSolvers}
\alias{getAvailableSolvers}
\title{Get the available solvers for use in trena}
\usage{
getAvailableSolvers()
}
\value{
A character vector of all solvers currently available in trena.
}
\description{
Retrieve the vector of different methods that can be used as solvers in trena. Solver names in
the returned vector correspond to the exact capitalization used in their Solver subclasses (i.e.
a Solver object using LASSO is LassoSolver)
}
\examples{
all.solvers <- getAvailableSolvers()

}
