% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFilter.R
\docType{methods}
\name{getCandidates,FootprintFilter-method}
\alias{getCandidates,FootprintFilter-method}
\alias{getCandidates-FootprintFilter}
\title{Get candidate genes using the variance filter}
\usage{
\S4method{getCandidates}{FootprintFilter}(obj)
}
\arguments{
\item{obj}{An object of class FootprintFilter}
}
\value{
A vector containing all genes with variances less than the target gene
}
\description{
Get candidate genes using the variance filter
}
\examples{

# Use footprint filter with the included SQLite database for MEF2C to filter candidates
# in the included Alzheimer's dataset
target.gene <- "MEF2C"
db.address <- system.file(package="trena", "extdata")
genome.db.uri <- paste("sqlite:/",db.address,"genome.sub.db", sep = "/")
project.db.uri <- paste("sqlite:/",db.address,"project.sub.db", sep = "/")
size.upstream <- 1000
size.downstream <- 1000
geneCenteredSpec <- list(targetGene = target.gene, tssUpstream = size.upstream, tssDownstream = size.downstream)
footprint.filter <- FootprintFilter(genomeDB = genome.db.uri, footprintDB = project.db.uri,
geneCenteredSpec = geneCenteredSpec)

tfs <- getCandidates(footprint.filter)
}
\seealso{
\code{\link{FootprintFilter}}

Other getCandidate Methods: \code{\link{getCandidates,VarianceFilter-method}},
  \code{\link{getCandidates}}
}
