% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneOntologyFilter.R
\docType{methods}
\name{getCandidates,GeneOntologyFilter-method}
\alias{getCandidates,GeneOntologyFilter-method}
\alias{getCandidates-GeneOntologyFilter}
\title{Get candidate genes using a gene ontology filter}
\usage{
\S4method{getCandidates}{GeneOntologyFilter}(obj)
}
\arguments{
\item{obj}{An object of class GeneOntologyFilter}
}
\value{
A list, where one element a character vector of transcription factors that match
the GO term and the other is an empty data frame.
}
\description{
Get candidate genes using a gene ontology filter
}
\examples{

# Make a filter for "transcription, DNA-templated" and use it to filter candidates
library(org.Hs.eg.db)
goFilter <- GeneOntologyFilter(org.Hs.eg.db, GOTerm="GO:0006351")
candidates <- getCandidates(goFilter)
}
\seealso{
\code{\link{GeneOntologyFilter}}

Other getCandidate Methods: \code{\link{getCandidates,FootprintFilter-method}},
  \code{\link{getCandidates,HumanDHSFilter-method}},
  \code{\link{getCandidates,VarianceFilter-method}},
  \code{\link{getCandidates}}
}
\concept{getCandidate Methods}
