% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullFilter.R
\docType{methods}
\name{getCandidates,NullFilter-method}
\alias{getCandidates,NullFilter-method}
\alias{getCandidates-NullFilter}
\title{Get candidate genes using the null filter}
\usage{
\S4method{getCandidates}{NullFilter}(obj, extraArgs = list())
}
\arguments{
\item{obj}{An object of class NullFilter}

\item{extraArgs}{An empty list}
}
\value{
A vector containing all genes in the assay matrix
}
\description{
Get candidate genes using the null filter
}
\examples{

# Using the included Alzheimer's data, return all transcription factors as candidates
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
null.filter <- NullFilter(mtx.assay=mtx.sub)
tfs <- getCandidates(null.filter)
}
\seealso{
\code{\link{NullFilter}}

Other getCandidate Methods: \code{\link{getCandidates,FootprintFilter-method}},
  \code{\link{getCandidates,VarianceFilter-method}},
  \code{\link{getCandidates}}
}
