% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarianceFilter.R
\docType{methods}
\name{getCandidates,VarianceFilter-method}
\alias{getCandidates,VarianceFilter-method}
\alias{getCandidates-VarianceFilter}
\title{Get candidate genes using the variance filter}
\usage{
\S4method{getCandidates}{VarianceFilter}(obj)
}
\arguments{
\item{obj}{An object of class VarianceFilter}
}
\value{
A vector containing all genes with variances less than the target gene
}
\description{
Get candidate genes using the variance filter
}
\examples{

# Using the included Alzheimer's dataset, filter out only those transcription factors with variance
# within 50\% of the variance of MEF2C
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
variance.filter <- VarianceFilter(mtx.assay = mtx.sub, targetGene = "MEF2C")
tfs <- getCandidates(variance.filter)
}
\seealso{
\code{\link{VarianceFilter}}

Other getCandidate Methods: \code{\link{getCandidates,FootprintFilter-method}},
  \code{\link{getCandidates,GeneOntologyFilter-method}},
  \code{\link{getCandidates,HumanDHSFilter-method}},
  \code{\link{getCandidates}}
}
\concept{getCandidate Methods}
