% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{getChromLoc,FootprintFinder-method}
\alias{getChromLoc,FootprintFinder-method}
\alias{getChromLoc}
\title{Get Chromasome Location}
\usage{
\S4method{getChromLoc}{FootprintFinder}(obj, name, biotype = "protein_coding",
  moleculetype = "gene")
}
\arguments{
\item{obj}{An object of class FootprintFinder}

\item{name}{A gene name or ID}

\item{biotype}{A type of biological unit (default="protein_coding")}

\item{moleculetype}{A type of molecule (default="gene")}
}
\value{
A dataframe containing the results of a database query pertaining to the specified name,
biotype, and molecule type. This dataframe contains the following columns: gene_id, gene_name,
chr, start, endpos, strand
}
\description{
Using the gtf table in the genome database contained in a FootprintFinder object, get the locations
of chromasomes with the specified gene name, biological unit type, and molecule type
}
\examples{
db.address <- system.file(package="TReNA", "extdata")
genome.db.uri <- paste("sqlite:/",db.address,"genome.sub.db", sep = "/")
project.db.uri <- paste("sqlite:/",db.address,"project.sub.db", sep = "/")
fp <- FootprintFinder(genome.db.uri, project.db.uri)

chrom.locs <- getChromLoc(fp, name = "MEF2C")
}
\seealso{
Other FootprintFinder methods: \code{\link{FootprintFinder-class}},
  \code{\link{closeDatabaseConnections,FootprintFinder-method}},
  \code{\link{getFootprintsForGene,FootprintFinder-method}},
  \code{\link{getFootprintsInRegion,FootprintFinder-method}},
  \code{\link{getGenePromoterRegion,FootprintFinder-method}},
  \code{\link{getGtfGeneBioTypes,FootprintFinder-method}},
  \code{\link{getGtfMoleculeTypes,FootprintFinder-method}},
  \code{\link{getPromoterRegionsAllGenes,FootprintFinder-method}},
  \code{\link{mapMotifsToTFsMergeIntoTable,FootprintFinder-method}}
}
