% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAMax.R
\name{getCoverage,PCAMax-method}
\alias{getCoverage,PCAMax-method}
\alias{getCoverage}
\title{what percentage of the variance is captured in the first two principal components?}
\usage{
\S4method{getCoverage}{PCAMax}(obj)
}
\arguments{
\item{obj}{An object of the class PCAMax}
}
\value{
a number between zero and one
}
\description{
what percentage of the variance is captured in the first two principal components?
}
