% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDHSFilter.R
\docType{methods}
\name{getEncodeRegulatoryTableNames,HumanDHSFilter-method}
\alias{getEncodeRegulatoryTableNames,HumanDHSFilter-method}
\alias{getEncodeRegulatoryTableNames}
\title{Get Encode regulatory tables using a human DHS filter}
\usage{
\S4method{getEncodeRegulatoryTableNames}{HumanDHSFilter}(obj)
}
\arguments{
\item{obj}{An object of class HumanDHSFilter}
}
\value{
A character vector containing the names of the Encode regulatory tables for the regions
contained in the HumanDHSFilter object
}
\description{
Get Encode regulatory tables using a human DHS filter
}
\examples{

#
}
\seealso{
\code{\link{HumanDHSFilter}}
}
