% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateFilter.R
\name{getFilterAssayData}
\alias{getFilterAssayData}
\title{Retrieve the assay matrix of gene expression data}
\usage{
getFilterAssayData(obj)
}
\arguments{
\item{obj}{An object of a CandidateFilter class}
}
\value{
The assay matrix of gene expression data associated with a CandidateFilter object
}
\description{
Retrieve the assay matrix of gene expression data
}
\examples{

# Create an empty CandidateFilter object
my.filter <- CandidateFilter()
}
