% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trena.R
\docType{methods}
\name{getGeneModelTableColumnNames,Trena-method}
\alias{getGeneModelTableColumnNames,Trena-method}
\alias{getGeneModelTableColumnNames}
\title{Retrieve the column names in the gene model table for a Trena object}
\usage{
\S4method{getGeneModelTableColumnNames}{Trena}(obj)
}
\arguments{
\item{obj}{An object of class Trena}
}
\value{
A character vector listing the column names in the Trena object gene model table
}
\description{
Retrieve the column names in the gene model table for a Trena object
}
\examples{
# Create a Trena object and retrieve the column names of the gene model table
trena <- Trena("mm10")
tbl.cols <- getRegulatoryTableColumnNames(trena)
}
