% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{getGenePromoterRegion,FootprintFinder-method}
\alias{getGenePromoterRegion,FootprintFinder-method}
\alias{getGenePromoterRegion}
\title{Get Gene Promoter Region}
\usage{
\S4method{getGenePromoterRegion}{FootprintFinder}(obj, gene,
  size.upstream = 1000, size.downstream = 0,
  biotype = "protein_coding", moleculetype = "gene")
}
\arguments{
\item{obj}{An object of class FootprintFinder}

\item{gene}{A gene name of ID}

\item{size.upstream}{An integer denoting the distance upstream of the target gene to look for footprints
(default = 1000)}

\item{size.downstream}{An integer denoting the distance downstream of the target gene to look for footprints
(default = 0)}

\item{biotype}{A type of biological unit (default="protein_coding")}

\item{moleculetype}{A type of molecule (default="gene")}
}
\value{
A list containing 3 elements:
1) chr : The name of the chromosome containing the promoter region for the specified gene
2) start : The starting location of the promoter region for the specified gene
3) end : The ending location of the promoter region for the specified gene
}
\description{
Using the \code{\link{getChromLoc}} function in conjunction with the gtf table inside the genome
database specified by the FootprintFinder object, get the chromosome, starting location,
and ending location for gene promoter region.
}
\examples{
db.address <- system.file(package="trena", "extdata")
genome.db.uri <- paste("sqlite:/",db.address,"mef2c.neighborhood.hg38.gtfAnnotation.db", sep = "/")
project.db.uri <- paste("sqlite:/",db.address,"mef2c.neigborhood.hg38.footprints.db", sep = "/")
fp <- FootprintFinder(genome.db.uri, project.db.uri)

prom.region <- getGenePromoterRegion(fp, gene = "MEF2C")
}
\seealso{
Other FootprintFinder methods: \code{\link{FootprintFinder-class}},
  \code{\link{closeDatabaseConnections,FootprintFinder-method}},
  \code{\link{getChromLoc,FootprintFinder-method}},
  \code{\link{getFootprintsForGene,FootprintFinder-method}},
  \code{\link{getFootprintsInRegion,FootprintFinder-method}},
  \code{\link{getGtfGeneBioTypes,FootprintFinder-method}},
  \code{\link{getGtfMoleculeTypes,FootprintFinder-method}},
  \code{\link{getPromoterRegionsAllGenes,FootprintFinder-method}}
}
\concept{FootprintFinder methods}
