% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifMatcher.R
\name{getPfms}
\alias{getPfms}
\title{Retrieve the motifs from the pfms slot}
\usage{
getPfms(obj)
}
\arguments{
\item{obj}{An object of class MotifMatcher}
}
\value{
The list of motif matrices stored in the pfms slot.
}
\description{
Given a MotifMatcher object, return the motifs, which are stored in the pfms slot.
}
\examples{

# Return the default matrix of JASPAR motifs
library(MotifDb)
motifMatcher <- MotifMatcher(genomeName="hg38", pfms = as.list(query(MotifDb, "sapiens")))
motifs <- getPfms(motifMatcher)
}
