% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{getPromoterRegionsAllGenes,FootprintFinder-method}
\alias{getPromoterRegionsAllGenes,FootprintFinder-method}
\alias{getPromoterRegionsAllGenes}
\title{Get Promoter Regions for All Genes}
\usage{
\S4method{getPromoterRegionsAllGenes}{FootprintFinder}(obj,
  size.upstream = 10000, size.downstream = 10000, use_gene_ids = TRUE)
}
\arguments{
\item{obj}{An object of class FootprintFinder}

\item{size.upstream}{An integer denoting the distance upstream of each gene's transcription start
site to include in the promoter region (default = 1000)}

\item{size.downstream}{An integer denoting the distance downstream of each gene's transcription start
site to include in the promoter region (default = 1000)}

\item{use_gene_ids}{A binary indicating whether to return gene IDs or gene names (default = T)}
}
\value{
A GRanges object containing the promoter regions for all genes
}
\description{
Using the gtf table inside the genome database specified by the FootprintFinder object, return the
promoter regions for every protein-coding gene in the database.
}
\examples{
db.address <- system.file(package="TReNA", "extdata")
genome.db.uri <- paste("sqlite:/",db.address,"genome.sub.db", sep = "/")
project.db.uri <- paste("sqlite:/",db.address,"project.sub.db", sep = "/")
fp <- FootprintFinder(genome.db.uri, project.db.uri)

footprints <- getPromoterRegionsAllGenes(fp)
}
\seealso{
Other FootprintFinder methods: \code{\link{FootprintFinder-class}},
  \code{\link{closeDatabaseConnections,FootprintFinder-method}},
  \code{\link{getChromLoc,FootprintFinder-method}},
  \code{\link{getFootprintsForGene,FootprintFinder-method}},
  \code{\link{getFootprintsInRegion,FootprintFinder-method}},
  \code{\link{getGenePromoterRegion,FootprintFinder-method}},
  \code{\link{getGtfGeneBioTypes,FootprintFinder-method}},
  \code{\link{getGtfMoleculeTypes,FootprintFinder-method}},
  \code{\link{mapMotifsToTFsMergeIntoTable,FootprintFinder-method}}
}
