% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{getRegulators,Solver-method}
\alias{getRegulators,Solver-method}
\alias{getRegulators}
\title{Retrieve the candiate regulators from a Solver object}
\usage{
\S4method{getRegulators}{Solver}(obj)
}
\arguments{
\item{obj}{An object of class Solver}
}
\value{
The candidate regulators associated with a Solver object
}
\description{
Retrieve the candiate regulators from a Solver object
}
\examples{
# Create a Solver object using the included Alzheimer's data and retrieve the regulators
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "MEF2C"
candidateRegulators <- setdiff(rownames(mtx.sub), targetGene)
solver <- Solver(mtx.sub, targetGene, candidateRegulators)
regs <- getRegulators(solver) 
}
