% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trena.R
\name{getRegulatoryTableColumnNames,Trena-method}
\alias{getRegulatoryTableColumnNames,Trena-method}
\alias{getRegulatoryTableColumnNames}
\title{Retrieve the column names in the regulatory table for a Trena object}
\usage{
\S4method{getRegulatoryTableColumnNames}{Trena}(obj)
}
\arguments{
\item{obj}{An object of class Trena}
}
\value{
A character vector listing the column names in the Trena object regulatory table
}
\description{
Retrieve the column names in the regulatory table for a Trena object
}
\examples{
# Create a Trena object and retrieve the column names of the regulatory table
trena <- Trena("mm10")
tbl.cols <- getRegulatoryTableColumnNames(trena)
}
