% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleSolver.R
\docType{methods}
\name{getSolverNames,EnsembleSolver-method}
\alias{getSolverNames,EnsembleSolver-method}
\alias{getSolverNames}
\title{Retrieve the solver names from an EnsembleSolver object}
\usage{
\S4method{getSolverNames}{EnsembleSolver}(obj)
}
\arguments{
\item{obj}{An object of class Solver}
}
\value{
The vector of solver names associated with an EnsembleSolver object
}
\description{
Retrieve the solver names from an EnsembleSolver object
}
\examples{
# Create a Solver object using the included Alzheimer's data and retrieve the regulators
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "MEF2C"
candidateRegulators <- setdiff(rownames(mtx.sub), targetGene)
solver <- EnsembleSolver(mtx.sub, targetGene, candidateRegulators,
solverNames = c("lasso","randomForest"))
solver.names <- getSolverNames(solver) 
}
