% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{getTarget}
\alias{getTarget}
\title{Retrieve the target gene from a Solver object}
\usage{
getTarget(obj)
}
\arguments{
\item{obj}{An object of class Solver}
}
\value{
The target gene associated with a Solver object
}
\description{
Retrieve the target gene from a Solver object
}
\examples{
# Create a Solver object using the included Alzheimer's data and retrieve the target gene
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "MEF2C"
candidateRegulators <- setdiff(rownames(mtx.sub), targetGene)
solver <- Solver(mtx.sub, targetGene, candidateRegulators)
mtx <- getTarget(solver) 
}
