% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoPVSolver.R
\docType{class}
\name{LassoPVSolver-class}
\alias{LassoPVSolver-class}
\alias{.LassoPVSolver}
\title{An S4 class to represent a LASSO P-Value solver}
\description{
An S4 class to represent a LASSO P-Value solver
}
