% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifMatcher.R
\docType{class}
\name{MotifMatcher-class}
\alias{MotifMatcher-class}
\alias{.MotifMatcher}
\alias{MotifMatcher}
\title{Create a MotifMatcher object}
\description{
A MotifMatcher object is used directly by the \code{\link{HumanDHSFilter}} class to match motif
matrices to where they occur in the supplied genome.
}
