% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAMax.R
\name{normalizeModel,PCAMax-method}
\alias{normalizeModel,PCAMax-method}
\alias{normalizeModel}
\title{transform a specific column to fit normal distribution}
\usage{
\S4method{normalizeModel}{PCAMax}(obj, normalizing.max = 10)
}
\arguments{
\item{obj}{An object of the class PCAMax}

\item{normalizing.max}{numeric, a maximum value for the normalized distrubtions}
}
\value{
a normalized matrix, each column treated separately
}
\description{
transform a specific column to fit normal distribution
}
