% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullFilter.R
\docType{class}
\name{NullFilter-class}
\alias{NullFilter-class}
\alias{.NullFilter}
\alias{NullFilter}
\alias{NullFilter}
\title{Create a NullFilter object}
\usage{
NullFilter(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the filter should print output}
}
\value{
A CandidateFilter class object with null as the filtering method

An object of the NullFilter class
}
\description{
A NullFilter object allows for "filtering" of the genes in an assay matrix. Its associated
\code{getCandidates} method returns the list of transcription factors included in the assay matrix.
}
\examples{
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
null.filter <- NullFilter(mtx.assay = mtx.sub)
}
\seealso{
\code{\link{getCandidates-NullFilter}}, \code{\link{getFilterAssayData}}

Other Filtering Objects: \code{\link{FootprintFilter-class}},
  \code{\link{VarianceFilter-class}}
}
