% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseChromLocString}
\alias{parseChromLocString}
\title{Parse a string containing a chromosome and location on the genome}
\usage{
parseChromLocString(chromLocString)
}
\arguments{
\item{chromLocString}{A string of the format "chromosome:start-end"}
}
\value{
A named list containing the chromosome, start, and end from the supplied string
}
\description{
Given a string of the format "chromosome:start-end", pull out the three values contained in
the string and return them as a named list
}
\examples{
# Note: Both examples return :
# list(chrom="chr10", start=118441047, end=118445892)

# Parse a string containing the "chr" prefix for chromosome
chrom.list <- parseChromLocString("chr10:118441047-118445892")

# Parse a string without the "chr" prefix in the chromosome
chrom.list <- parseChromLocString("10:118441047-118445892")

}
