% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseDatabaseUri}
\alias{parseDatabaseUri}
\title{Parse a string containing the information for connecting to a database}
\usage{
parseDatabaseUri(database.uri)
}
\arguments{
\item{database.uri}{A string of the format "database_driver://host/database_name"}
}
\value{
A named list containing the driver, host, and database name from the
supplied string
}
\description{
Given a string of the format "database_driver://host/database_name", pull out the 3 pieces of
information and return them as a named list.
}
\examples{

# Parse a URI for a local PostgreSQL database called "gtf"
parsed.URI <- parseDatabaseUri("postgres://localhost/gtf")

# Parse a URI for the included SQLite database "vrk2.genome.db" in the local documents folder
db.address <- system.file(package="trena", "extdata")
genome.db.uri    <- paste("sqlite:/", db.address, "vrk2.genome.db",  sep = "/")
parsed.URI <- parseDatabaseUri(genome.db.uri)

}
