% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAMax.R
\docType{class}
\name{PCAMax}
\alias{PCAMax}
\alias{.pcaMax}
\title{Class PCAMax}
\usage{
PCAMax(tbl, tfIdentifierColumnName = "tf.hgnc")
}
\arguments{
\item{tbl}{a data.frame}
}
\value{
a PCAMax object
}
\description{
Class PCAMax

Create a PCAMax object from a data.frame produced by the EnsembleSolver
}
\seealso{
\code{\link{EnsembleSolver}}
}
