% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearsonSolver.R
\name{PearsonSolver}
\alias{PearsonSolver}
\title{Create a Solver class object using  Pearson correlation coefficients as the solver}
\usage{
PearsonSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with Pearson correlation coefficients as the solver
}
\description{
Create a Solver class object using  Pearson correlation coefficients as the solver
}
\examples{
solver <- PearsonSolver()
}
\seealso{
\code{\link{solve.Pearson}}, \code{\link{getAssayData}}

Other Solver class objects: \code{\link{BayesSpikeSolver}},
  \code{\link{EnsembleSolver}},
  \code{\link{LassoPVSolver}}, \code{\link{LassoSolver}},
  \code{\link{RandomForestSolver}},
  \code{\link{RidgeSolver}}, \code{\link{Solver-class}},
  \code{\link{SpearmanSolver}},
  \code{\link{SqrtLassoSolver}}
}
