% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSolver.R
\name{RandomForestSolver}
\alias{RandomForestSolver}
\title{Create a Solver class object using the Random Forest solver}
\usage{
RandomForestSolver(
  mtx.assay = matrix(),
  targetGene,
  candidateRegulators,
  regulatorWeights = rep(1, length(candidateRegulators)),
  quiet = TRUE
)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{targetGene}{A designated target gene that should be part of the mtx.assay data}

\item{candidateRegulators}{The designated set of transcription factors that could be associated
with the target gene}

\item{regulatorWeights}{A set of weights on the transcription factors
(default = rep(1, length(candidateRegulators)))}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with Random Forest as the solver
}
\description{
Create a Solver class object using the Random Forest solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "MEF2C"
candidateRegulators <- setdiff(rownames(mtx.sub), targetGene)
rf.solver <- RandomForestSolver(mtx.sub, targetGene, candidateRegulators)
}
\seealso{
\code{\link{solve.RandomForest}}, \code{\link{getAssayData}}

Other Solver class objects: 
\code{\link{BicorSolver}},
\code{\link{EnsembleSolver}},
\code{\link{HumanDHSFilter-class}},
\code{\link{LassoPVSolver}},
\code{\link{LassoSolver}},
\code{\link{PearsonSolver}},
\code{\link{RidgeSolver}},
\code{\link{Solver-class}},
\code{\link{SpearmanSolver}},
\code{\link{XGBoostSolver}}
}
\concept{Solver class objects}
