% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\docType{methods}
\name{rescalePredictorWeights,Solver-method}
\alias{rescalePredictorWeights,Solver-method}
\alias{rescalePredictorWeights}
\title{Rescale the Predictor Weights}
\usage{
\S4method{rescalePredictorWeights}{Solver}(obj, rawValue.min, rawValue.max,
  rawValues)
}
\arguments{
\item{obj}{An object of the Solver class}

\item{rawValue.min}{The minimum value of the raw expression values}

\item{rawValue.max}{The maximum value of the raw expression values}

\item{rawValues}{A matrix of raw expression values}
}
\value{
A matrix of the raw values re-scaled using the minimum and maximum values
}
\description{
Solvers such as LASSO penalize predictors on a scale of 1 (full weight) to infinity (zero weight).
With the \code{rescalePredictorWeights} method, incoming raw values can be scaled between a possibly
theoretical minimum and maximum value.
}
\examples{
# Create a LassoSolver object using the included Alzheimer's data and rescale the predictors
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
ls <- LassoSolver(mtx.sub)
raw.values <- c(241, 4739, 9854, 22215, 658334)
cooked.values <- rescalePredictorWeights(ls, rawValue.min = 1, rawValue.max = 1000000, raw.values)
}
