% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeSolver.R
\name{RidgeSolver}
\alias{RidgeSolver}
\title{Create a Solver class object using the Ridge solver}
\usage{
RidgeSolver(
  mtx.assay = matrix(),
  targetGene,
  candidateRegulators,
  regulatorWeights = rep(1, length(candidateRegulators)),
  alpha = 0,
  lambda = numeric(0),
  keep.metrics = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{targetGene}{A designated target gene that should be part of the mtx.assay data}

\item{candidateRegulators}{The designated set of transcription factors that could be associated
with the target gene}

\item{regulatorWeights}{A set of weights on the transcription factors
(default = rep(1, length(tfs)))}

\item{alpha}{A parameter from 0-1 that determines the proportion of LASSO to ridge used in the
elastic net solver, with 0 being fully ridge and 1 being fully LASSO (default = 0.9)}

\item{lambda}{A tuning parameter that determines the severity of the penalty function imposed
on the elastic net regression. If unspecified, lambda will be determined via
permutation testing (default = numeric(0)).}

\item{keep.metrics}{A logical denoting whether or not to keep the initial supplied metrics
versus determining new ones}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with Ridge as the solver
}
\description{
Create a Solver class object using the Ridge solver
}
\examples{
\dontrun{
   load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
   target.gene <- "MEF2C"
   tfs <- setdiff(rownames(mtx.sub), target.gene)
   ridge.solver <- RidgeSolver(mtx.sub, target.gene, tfs)
   }
}
\seealso{
\code{\link{solve.Ridge}}, \code{\link{getAssayData}}

Other Solver class objects: 
\code{\link{BicorSolver}},
\code{\link{EnsembleSolver}},
\code{\link{HumanDHSFilter-class}},
\code{\link{LassoPVSolver}},
\code{\link{LassoSolver}},
\code{\link{PearsonSolver}},
\code{\link{RandomForestSolver}},
\code{\link{Solver-class}},
\code{\link{SpearmanSolver}},
\code{\link{XGBoostSolver}}
}
\concept{Solver class objects}
