% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{run}
\alias{run}
\title{Run a  Solver object to select features}
\usage{
run(obj)
}
\arguments{
\item{obj}{An object of a Solver class}
}
\value{
A data frame of candidate regulators and the relation to the target gene
}
\description{
Run a  Solver object to select features
}
