% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpikeSolver.R
\name{show,BayesSpikeSolver-method}
\alias{show,BayesSpikeSolver-method}
\alias{show.BayesSpikeSolver}
\title{Show the Bayes Spike Solver}
\usage{
\S4method{show}{BayesSpikeSolver}(object)
}
\arguments{
\item{object}{An object of the class BayesSpikeSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Bayes Spike Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
bayes.solver <- BayesSpikeSolver(mtx.sub, target.gene, tfs)
show(bayes.solver)
}
