% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicorSolver.R
\name{show,BicorSolver-method}
\alias{show,BicorSolver-method}
\alias{show.BicorSolver}
\title{Show the Bicor Solver}
\usage{
\S4method{show}{BicorSolver}(object)
}
\arguments{
\item{object}{An object of the class BicorSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Bicor Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
bicor.solver <- BicorSolver(mtx.sub, target.gene, tfs)
show(bicor.solver)
}
