% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleSolver.R
\docType{methods}
\name{show,EnsembleSolver-method}
\alias{show,EnsembleSolver-method}
\alias{show.EnsembleSolver}
\title{Show the Ensemble Solver}
\usage{
\S4method{show}{EnsembleSolver}(object)
}
\arguments{
\item{object}{An object of the class EnsembleSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Ensemble Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
ensemble.solver <- EnsembleSolver(mtx.sub, target.gene, tfs)
show(ensemble.solver)
}
