% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoPVSolver.R
\docType{methods}
\name{show,LassoPVSolver-method}
\alias{show,LassoPVSolver-method}
\alias{show.LassoPVSolver}
\title{Show the Lasso PV Solver}
\usage{
\S4method{show}{LassoPVSolver}(obj)
}
\arguments{
\item{obj}{An object of the class LassoPVSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Lasso PV Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
lassopv.solver <- LassoPVSolver(mtx.sub, target.gene, tfs)
show(lassopv.solver)
}
