% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoSolver.R
\docType{methods}
\name{show,LassoSolver-method}
\alias{show,LassoSolver-method}
\alias{show.LassoSolver}
\title{Show the Lasso Solver}
\usage{
\S4method{show}{LassoSolver}(obj)
}
\arguments{
\item{obj}{An object of the class LassoSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Lasso Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
lasso.solver <- LassoSolver(mtx.sub, target.gene, tfs)
show(lasso.solver)
}
