% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifMatcher.R
\docType{methods}
\name{show,MotifMatcher-method}
\alias{show,MotifMatcher-method}
\alias{show.MotifMatcher}
\title{Show a MotifMatcher object}
\usage{
\S4method{show}{MotifMatcher}(object)
}
\arguments{
\item{object}{An object of the class MotifMatcher}
}
\value{
A truncated view of the supplied object
}
\description{
Show a MotifMatcher object
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
lassopv.solver <- LassoPVSolver(mtx.sub, target.gene, tfs)
show(lassopv.solver)
}
