% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearsonSolver.R
\docType{methods}
\name{show,PearsonSolver-method}
\alias{show,PearsonSolver-method}
\alias{show.PearsonSolver}
\title{Show the Pearson Solver}
\usage{
\S4method{show}{PearsonSolver}(object)
}
\arguments{
\item{object}{An object of the class PearsonSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Pearson Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
pearson.solver <- PearsonSolver(mtx.sub, target.gene, tfs)
show(pearson.solver)
}
