% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSolver.R
\docType{methods}
\name{show,RandomForestSolver-method}
\alias{show,RandomForestSolver-method}
\alias{show.RandomForestSolver}
\title{Show the Random Forest Solver}
\usage{
\S4method{show}{RandomForestSolver}(object)
}
\arguments{
\item{object}{An object of the class RandomForestSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Random Forest Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
rf.solver <- RandomForestSolver(mtx.sub, target.gene, tfs)
show(rf.solver)
}
