% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{show}
\alias{show}
\title{Show a Solver object}
\usage{
show(obj)
}
\arguments{
\item{obj}{An object of class Solver}
}
\value{
The names of the top 10 candidate regulators, plus statistics on the assay matrix, the
target gene, and the total number of candidate regulators
}
\description{
Show a Solver object
}
\examples{
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
tfs <- setdiff(rownames(mtx.sub),"MEF2C")
pearson.solver <- PearsonSolver(mtx.sub, "MEF2C", tfs)
show(pearson.solver)
}
