% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpearmanSolver.R
\name{show,SpearmanSolver-method}
\alias{show,SpearmanSolver-method}
\alias{show.SpearmanSolver}
\title{Show the Spearman Solver}
\usage{
\S4method{show}{SpearmanSolver}(object)
}
\arguments{
\item{object}{An object of the class SpearmanSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Spearman Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
spearman.solver <- SpearmanSolver(mtx.sub, target.gene, tfs)
show(spearman.solver)
}
