% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\docType{methods}
\name{show,SqrtLassoSolver-method}
\alias{show,SqrtLassoSolver-method}
\alias{show.SqrtLassoSolver}
\title{Show the Square Root Lasso Solver}
\usage{
\S4method{show}{SqrtLassoSolver}(object)
}
\arguments{
\item{object}{An object of the class SqrtLassoSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Square Root Lasso Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
sqrt.solver <- SqrtLassoSolver(mtx.sub, target.gene, tfs)
show(sqrt.solver)
}
