% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostSolver.R
\docType{methods}
\name{show,XGBoostSolver-method}
\alias{show,XGBoostSolver-method}
\alias{show.XGBoostSolver}
\title{Show the XGBoost Solver}
\usage{
\S4method{show}{XGBoostSolver}(object)
}
\arguments{
\item{object}{An object of the class XGBoostSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the XGBoost Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
XGBoost.solver <- XGBoostSolver(mtx.sub, target.gene, tfs)
show(XGBoost.solver)
}
