% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpikeSolver.R
\docType{methods}
\name{run,BayesSpikeSolver-method}
\alias{run,BayesSpikeSolver-method}
\alias{run.BayesSpikeSolver}
\alias{solve.BayesSpike}
\title{Run the Bayes Spike Solver}
\usage{
\S4method{run}{BayesSpikeSolver}(obj)
}
\arguments{
\item{obj}{An object of the class BayesSpikeSolver}
}
\value{
A data frame containing the coefficients relating the target gene to each transcription factor,
plus other fit parameters
}
\description{
Given a TReNA object with Bayes Spike as the solver, use the \code{\link{vbsr}}
function to estimate coefficients for each transcription factor as a predictor of the target
gene's expression level.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
bayes.solver <- BayesSpikeSolver(mtx.sub, target.gene, tfs)
tbl <- run(bayes.solver)

# Solve the same Alzheimer's problem, but this time set the number of random starts to 100
bayes.solver <- BayesSpikeSolver(mtx.sub, target.gene, tfs, nOrderings = 100)
tbl <- run(bayes.solver)
}
\seealso{
\code{\link{vbsr}}, \code{\link{BayesSpikeSolver}}

Other solver methods: \code{\link{run,EnsembleSolver-method}},
  \code{\link{run,LassoPVSolver-method}},
  \code{\link{run,LassoSolver-method}},
  \code{\link{run,PearsonSolver-method}},
  \code{\link{run,RandomForestSolver-method}},
  \code{\link{run,RidgeSolver-method}},
  \code{\link{run,SpearmanSolver-method}},
  \code{\link{run,SqrtLassoSolver-method}},
  \code{\link{solve,TReNA-method}}
}
