% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicorSolver.R
\name{run,BicorSolver-method}
\alias{run,BicorSolver-method}
\alias{run.BicorSolver}
\alias{solve.Bicor}
\title{Run the Bicor Solver}
\usage{
\S4method{run}{BicorSolver}(obj)
}
\arguments{
\item{obj}{An object of class BicorSolver}
}
\value{
The set of Bicor Correlation Coefficients between each transcription factor and the target gene.
}
\description{
Given a BicorSolver object, use the \code{\link{cor}} function
to estimate coefficients for each transcription factor as a predictor of the target gene's
expression level.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
bicor.solver <- BicorSolver(mtx.sub, target.gene, tfs)
tbl <- run(bicor.solver)
}
\seealso{
\code{\link{cor}}, \code{\link{BicorSolver}}

Other solver methods: 
\code{\link{run,EnsembleSolver-method}},
\code{\link{run,LassoPVSolver-method}},
\code{\link{run,LassoSolver-method}},
\code{\link{run,PearsonSolver-method}},
\code{\link{run,RandomForestSolver-method}},
\code{\link{run,RidgeSolver-method}},
\code{\link{run,SpearmanSolver-method}},
\code{\link{run,XGBoostSolver-method}}
}
\concept{solver methods}
