% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleSolver.R
\docType{methods}
\name{run,EnsembleSolver-method}
\alias{run,EnsembleSolver-method}
\alias{run.EnsembleSolver}
\alias{solve.Ensemble}
\title{Run the Ensemble Solver}
\usage{
\S4method{run}{EnsembleSolver}(obj)
}
\arguments{
\item{obj}{An object of class Solver with "ensemble" as the solver string}
}
\value{
A data frame containing the scores for all solvers and two composite scores
relating the target gene to each transcription factor. The two new scores are:
\itemize{
\item{"concordance": a composite score created similarly to "extreme_score", but with each solver's
score scaled using *atan(x)*. This score scales from 0-1}
\item{"pcaMax": a composite score created using the root mean square of the principal
components of the individual solver scores}
}
}
\description{
Given a TReNA object with Ensemble as the solver and a list of solvers
(default = "default.solvers"), estimate coefficients for each transcription factor
as a predictor of the target gene's expression level. The final scores for the ensemble
method combine all specified solvers to create a composite score for each transcription factor.
This method should be called using the \code{\link{solve}} method on an appropriate TReNA object.
}
\seealso{
\code{\link{EnsembleSolver}}

Other solver methods: \code{\link{run,BayesSpikeSolver-method}},
  \code{\link{run,LassoPVSolver-method}},
  \code{\link{run,LassoSolver-method}},
  \code{\link{run,PearsonSolver-method}},
  \code{\link{run,RandomForestSolver-method}},
  \code{\link{run,RidgeSolver-method}},
  \code{\link{run,SpearmanSolver-method}},
  \code{\link{run,SqrtLassoSolver-method}}
}
\concept{solver methods}
