% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearsonSolver.R
\name{run,PearsonSolver-method}
\alias{run,PearsonSolver-method}
\alias{run.PearsonSolver}
\alias{solve.Pearson}
\title{Run the Pearson Solver}
\usage{
\S4method{run}{PearsonSolver}(obj)
}
\arguments{
\item{obj}{An object of class PearsonSolver}
}
\value{
The set of Pearson Correlation Coefficients between each transcription factor and the target gene.
}
\description{
Given a PearsonSolver object, use the \code{\link{cor}} function
to estimate coefficients for each transcription factor as a predictor of the target gene's
expression level.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
pearson.solver <- PearsonSolver(mtx.sub, target.gene, tfs)
tbl <- run(pearson.solver)
}
\seealso{
\code{\link{cor}}, \code{\link{PearsonSolver}}

Other solver methods: 
\code{\link{run,BicorSolver-method}},
\code{\link{run,EnsembleSolver-method}},
\code{\link{run,LassoPVSolver-method}},
\code{\link{run,LassoSolver-method}},
\code{\link{run,RandomForestSolver-method}},
\code{\link{run,RidgeSolver-method}},
\code{\link{run,SpearmanSolver-method}},
\code{\link{run,XGBoostSolver-method}}
}
\concept{solver methods}
