% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSolver.R
\name{run,RandomForestSolver-method}
\alias{run,RandomForestSolver-method}
\alias{run.RandomForestSolver}
\alias{solve.RandomForest}
\title{Run the Random Forest Solver}
\usage{
\S4method{run}{RandomForestSolver}(obj)
}
\arguments{
\item{obj}{An object of class TReNA with "randomForest" as the solver string}
}
\value{
A data frame containing the IncNodePurity for each candidate regulator. This coefficient
estimates the relationship between the candidates and the target gene.
}
\description{
Given a TReNA object with RandomForest as the solver, use the \code{\link{randomForest}} function
to estimate coefficients for each transcription factor as a predictor of the target gene's
expression level.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Random Forest as solver, and solve
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "MEF2C"
candidateRegulators <- setdiff(rownames(mtx.sub), targetGene)
rf.solver <- RandomForestSolver(mtx.sub, targetGene, candidateRegulators)
tbl <- run(rf.solver)
}
\seealso{
\code{\link{randomForest}}, \code{\link{RandomForestSolver}}

Other solver methods: 
\code{\link{run,BayesSpikeSolver-method}},
\code{\link{run,EnsembleSolver-method}},
\code{\link{run,LassoPVSolver-method}},
\code{\link{run,LassoSolver-method}},
\code{\link{run,PearsonSolver-method}},
\code{\link{run,RidgeSolver-method}},
\code{\link{run,SpearmanSolver-method}},
\code{\link{run,XGBoostSolver-method}}
}
\concept{solver methods}
