% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TReNA.R
\docType{methods}
\name{solve,TReNA-method}
\alias{solve,TReNA-method}
\alias{solve}
\alias{solve.TReNA}
\title{Solve a TReNA Object}
\usage{
\S4method{solve}{TReNA}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs)), extraArgs = list())
}
\arguments{
\item{obj}{An object of class TReNA}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}

\item{extraArgs}{Modifiers to the solver}
}
\value{
A data frame containing coefficients relating the target gene to each transcription factor
}
\description{
A TReNA object contains an assay matrix with expression data for genes of interest and a string
representing the chosen solver. The \code{solve} method runs the specified solver given a target
gene and a designated set of transcription factors, returning a list of parameters that quantify
the relationship between the transcription factors and the target gene.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with LASSO as the solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
trena <- TReNA(mtx.assay = mtx.sub, solver = "lasso")
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
tbl <- solve(trena, target.gene, tfs)
}
\seealso{
\code{\link{TReNA}}

Other solver methods: \code{\link{run,BayesSpikeSolver-method}},
  \code{\link{run,EnsembleSolver-method}},
  \code{\link{run,LassoPVSolver-method}},
  \code{\link{run,LassoSolver-method}},
  \code{\link{run,PearsonSolver-method}},
  \code{\link{run,RandomForestSolver-method}},
  \code{\link{run,RidgeSolver-method}},
  \code{\link{run,SpearmanSolver-method}},
  \code{\link{run,SqrtLassoSolver-method}}
}
