% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeSolver.R
\docType{methods}
\name{run,RidgeSolver-method}
\alias{run,RidgeSolver-method}
\alias{run.RidgeSolver}
\alias{solve.Ridge}
\title{Run the Ridge Regression Solver}
\usage{
\S4method{run}{RidgeSolver}(obj)
}
\arguments{
\item{obj}{An object of class RidgeSolver}
}
\value{
A data frame containing the coefficients relating the target gene to each
transcription factor, plus other fit parameters.
}
\description{
Given a TReNA object with Ridge Regression as the solver,
use the \code{\link{glmnet}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
ridge.solver <- LassoSolver(mtx.sub, target.gene, tfs)
tbl <- run(ridge.solver)
}
\seealso{
\code{\link{glmnet}}, , \code{\link{RidgeSolver}}

Other solver methods: \code{\link{run,BayesSpikeSolver-method}},
  \code{\link{run,EnsembleSolver-method}},
  \code{\link{run,LassoPVSolver-method}},
  \code{\link{run,LassoSolver-method}},
  \code{\link{run,PearsonSolver-method}},
  \code{\link{run,RandomForestSolver-method}},
  \code{\link{run,SpearmanSolver-method}},
  \code{\link{run,SqrtLassoSolver-method}},
  \code{\link{solve,TReNA-method}}
}
