% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpearmanSolver.R
\name{run,SpearmanSolver-method}
\alias{run,SpearmanSolver-method}
\alias{run.SpearmanSolver}
\alias{solve.Spearman}
\title{Run the Spearman Solver}
\usage{
\S4method{run}{SpearmanSolver}(obj)
}
\arguments{
\item{obj}{An object of class SpearmanSolver}
}
\value{
The set of Spearman Correlation Coefficients between each transcription factor and the target gene.
}
\description{
Given a TReNA object with Spearman as the solver, use the \code{\link{cor}}
function with \code{method = "spearman"} to esimate coefficients for each transcription factor
as a predictor of the target gene's expression level.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
spearman.solver <- SpearmanSolver(mtx.sub, target.gene, tfs)
tbl <- run(spearman.solver)
}
\seealso{
\code{\link{cor}}, \code{\link{SpearmanSolver}}

Other solver methods: 
\code{\link{run,BayesSpikeSolver-method}},
\code{\link{run,EnsembleSolver-method}},
\code{\link{run,LassoPVSolver-method}},
\code{\link{run,LassoSolver-method}},
\code{\link{run,PearsonSolver-method}},
\code{\link{run,RandomForestSolver-method}},
\code{\link{run,RidgeSolver-method}},
\code{\link{run,XGBoostSolver-method}}
}
\concept{solver methods}
