% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\docType{methods}
\name{run,SqrtLassoSolver-method}
\alias{run,SqrtLassoSolver-method}
\alias{run.SqrtLassoSolver}
\alias{solve.SqrtLasso}
\title{Run the Square Root LASSO Solver}
\usage{
\S4method{run}{SqrtLassoSolver}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs)), extraArgs = list())
}
\arguments{
\item{obj}{An object of class Solver with "sqrtlasso" as the solver string}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}

\item{extraArgs}{Modifiers to the Square Root LASSO solver}
}
\value{
A data frame containing the coefficients relating the target gene to
each transcription factor, plus other fit parameters.
}
\description{
Given a TReNA object with Square Root LASSO as the solver,
use the \code{\link{slim}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
This method should be called using the \code{\link{solve}} method on an appropriate TReNA object.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Square Root LASSO as solver, and solve
# Use 4 cores with the extraArgs argument
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
trena <- TReNA(mtx.assay = mtx.sub, solver = "sqrtlasso")
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
tbl <- solve(trena, target.gene, tfs, extraArgs = list("num.cores" = 4))
}
\seealso{
\code{\link{slim}}, \code{\link{SqrtLassoSolver}}

Other solver methods: \code{\link{run,BayesSpikeSolver-method}},
  \code{\link{run,EnsembleSolver-method}},
  \code{\link{run,LassoPVSolver-method}},
  \code{\link{run,LassoSolver-method}},
  \code{\link{run,PearsonSolver-method}},
  \code{\link{run,RandomForestSolver-method}},
  \code{\link{run,RidgeSolver-method}},
  \code{\link{run,SpearmanSolver-method}},
  \code{\link{solve,TReNA-method}}
}
