% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostSolver.R
\name{run,XGBoostSolver-method}
\alias{run,XGBoostSolver-method}
\alias{run.XGBoostSolver}
\alias{solve.XGBoost}
\title{Run the XGBoost Solver}
\usage{
\S4method{run}{XGBoostSolver}(obj)
}
\arguments{
\item{obj}{An object of class XGBoostSolver}
}
\value{
The set of XGBoost relative importances between each transcription factor and the target gene.
}
\description{
Given a TReNA object with XGBoost as the solver, use the \code{\link{cor}}
function with \code{method = "XGBoost"} to esimate importances for each transcription factor
as a predictor of the target gene's expression level.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
XGBoost.solver <- XGBoostSolver(mtx.sub, target.gene, tfs)
tbl <- run(XGBoost.solver)
}
\seealso{
\code{\link{cor}}, \code{\link{XGBoostSolver}}

Other solver methods: 
\code{\link{run,BicorSolver-method}},
\code{\link{run,EnsembleSolver-method}},
\code{\link{run,LassoPVSolver-method}},
\code{\link{run,LassoSolver-method}},
\code{\link{run,PearsonSolver-method}},
\code{\link{run,RandomForestSolver-method}},
\code{\link{run,RidgeSolver-method}},
\code{\link{run,SpearmanSolver-method}}
}
\concept{solver methods}
