% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\docType{class}
\name{Solver-class}
\alias{Solver-class}
\alias{.Solver}
\alias{Solver}
\alias{Solver}
\title{Define an object of class Solver}
\usage{
Solver(mtx.assay = matrix(), targetGene, candidateRegulators, quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{targetGene}{A designated target gene that should be part of the mtx.assay data}

\item{candidateRegulators}{The designated set of transcription factors that could be associated}

\item{quiet}{A logical indicating whether or not the Solver object should print output}
}
\value{
An object of the Solver class
}
\description{
The Solver class is a generic class that governs the different solvers available in TReNA. A
Solver class object is constructed during creation of a TReNA object and resides within the
TReNA object. It is rarely called by itself; rather, interaction with a particular solver object
is achieved using the \code{\link{solve}} method on a TReNA object.
}
\examples{
# Create a simple Solver object with default options
mtx <- matrix(rnorm(10000), nrow = 100)
solver <- Solver(mtx)

}
\seealso{
\code{\link{getAssayData}}, \code{\link{getTarget}}, \code{\link{getRegulators}}

Other Solver class objects: \code{\link{BayesSpikeSolver}},
  \code{\link{EnsembleSolver}},
  \code{\link{GeneOntologyFilter}},
  \code{\link{HumanDHSFilter}},
  \code{\link{LassoPVSolver}}, \code{\link{LassoSolver}},
  \code{\link{PearsonSolver}},
  \code{\link{RandomForestSolver}},
  \code{\link{RidgeSolver}}, \code{\link{SpearmanSolver}},
  \code{\link{SqrtLassoSolver}}
}
