% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpearmanSolver.R
\docType{class}
\name{SpearmanSolver-class}
\alias{SpearmanSolver-class}
\alias{.SpearmanSolver}
\title{An S4 class to represent a Spearman solver}
\description{
An S4 class to represent a Spearman solver
}
