% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpearmanSolver.R
\name{SpearmanSolver}
\alias{SpearmanSolver}
\title{Create a Solver class object using Spearman correlation coefficients as the solver}
\usage{
SpearmanSolver(mtx.assay = matrix(), targetGene, candidateRegulators,
  quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{candidateRegulators}{The designated set of transcription factors that could be associated
with the target gene}

\item{quiet}{A logical denoting whether or not the solver should print output}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}
}
\value{
A Solver class object with Spearman correlation coefficients as the solver
}
\description{
Create a Solver class object using Spearman correlation coefficients as the solver
}
\examples{
solver <- SpearmanSolver()
}
\seealso{
\code{\link{solve.Spearman}}, \code{\link{getAssayData}}

Other Solver class objects: \code{\link{BayesSpikeSolver}},
  \code{\link{EnsembleSolver}},
  \code{\link{LassoPVSolver}}, \code{\link{LassoSolver}},
  \code{\link{PearsonSolver}},
  \code{\link{RandomForestSolver}},
  \code{\link{RidgeSolver}}, \code{\link{Solver-class}},
  \code{\link{SqrtLassoSolver}}
}
