% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\docType{class}
\name{SqrtLassoSolver-class}
\alias{SqrtLassoSolver-class}
\alias{.SqrtLassoSolver}
\title{An S4 class to represent a Square Root LASSO solver}
\description{
An S4 class to represent a Square Root LASSO solver
}
