% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trena.R
\docType{class}
\name{Trena-class}
\alias{Trena-class}
\alias{.Trena}
\alias{Trena}
\title{Define an object of class Trena}
\usage{
Trena(genomeName, quiet = TRUE)
}
\arguments{
\item{genomeName}{A string indicating the genome used by the Trena object. Currently, only human and mouse (
"hg38","mm10") are supported}

\item{quiet}{A logical indicating whether or not the Trena object should print output}
}
\value{
An object of the Trena class
}
\description{
The Trena class provides a convenient wrapper for the most commonly used filters and solvers in the \code{trena}
package. Given a particular genome (one of \code{c("hg38","mm10")}, the Trena class provides methods to
retrieve information about possible regulators for a target gene, assess the effects of SNPs, and create gene models
using the flexible \code{\link{EnsembleSolver}} class.
}
\examples{
# Create a Trena object using the human hg38 genome
trena <- Trena("hg38")

}
\seealso{
\code{\link{getRegulatoryChromosomalRegions}}, \code{\link{getRegulatoryTableColumnNames}},
\code{\link{getGeneModelTableColumnNames}},
\code{\link{createGeneModelFromRegulatoryRegions}},
\code{\link{createGeneModelFromTfList}}
}
