% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoostSolver.R
\name{XGBoostSolver}
\alias{XGBoostSolver}
\title{Create a Solver class using gradient boosting (a regression technique) and the XGBoost library}
\usage{
XGBoostSolver(
  mtx.assay = matrix(),
  targetGene,
  candidateRegulators,
  quiet = TRUE
)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{targetGene}{A designated target gene that should be part of the mtx.assay data}

\item{candidateRegulators}{The designated set of transcription factors that could be associated
with the target gene}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with XGBoost Importances (Gain) as the solver
}
\description{
Create a Solver class using gradient boosting (a regression technique) and the XGBoost library
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
XGBoost.solver <- XGBoostSolver(mtx.sub, target.gene, tfs)
}
\seealso{
\code{\link{solve.XGBoost}}, \code{\link{getAssayData}}

Other Solver class objects: 
\code{\link{BayesSpikeSolver}},
\code{\link{EnsembleSolver}},
\code{\link{HumanDHSFilter-class}},
\code{\link{LassoPVSolver}},
\code{\link{LassoSolver}},
\code{\link{PearsonSolver}},
\code{\link{RandomForestSolver}},
\code{\link{RidgeSolver}},
\code{\link{Solver-class}},
\code{\link{SpearmanSolver}}
}
\concept{Solver class objects}
