% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trena.R
\name{assessSnp}
\alias{assessSnp}
\title{Assess the effect of a SNP using a Trena object}
\usage{
assessSnp(obj, pfms, variant, shoulder, pwmMatchMinimumAsPercentage,
  genomeName = "hg38")
}
\arguments{
\item{obj}{An object of class Trena}

\item{pfms}{A set of motif matrices, generally retrieved using MotifDb}

\item{variant}{A variant of interest}

\item{shoulder}{A distance from the TSS to use as a window}

\item{pwmMatchMinimumAsPercentage}{A minimum match percentage for the  motifs}

\item{genomeName}{A genome name (one of "hg38", "hg19", "mm10")}
}
\value{
A data frame containing the gene model
}
\description{
Assess the effect of a SNP using a Trena object
}
\examples{
# Create a Trena object for human, retrieve regions for "MEF2C" using a footprint filter,
# and then add the TF column
trena <- Trena("hg38")
chromosome <- "chr5"
mef2c.tss <- 88904257
loc.start <- mef2c.tss - 1000
loc.end   <- mef2c.tss + 1000

regions <- getRegulatoryChromosomalRegions(trena, chromosome, mef2c.tss-1000, mef2c.tss+1000,
sources, "MEF2C", mef2c.tss)
regions.tf <- expandRegulatoryRegionsTableByTF(trena, regions)
}
