% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trena.R
\name{assessSnp}
\alias{assessSnp}
\title{Assess the effect of a SNP using a Trena object}
\usage{
assessSnp(obj, pfms, variant, shoulder, pwmMatchMinimumAsPercentage,
  relaxedMatchDelta = 25)
}
\arguments{
\item{obj}{An object of class Trena}

\item{pfms}{A set of motif matrices, generally retrieved using MotifDb}

\item{variant}{A variant of interest}

\item{shoulder}{A distance from the TSS to use as a window}

\item{pwmMatchMinimumAsPercentage}{A minimum match percentage for the  motifs}

\item{relaxedMatchDelta}{A numeric indicating the degree of the match (default = 25)}
}
\value{
A data frame containing the gene model
}
\description{
Assess the effect of a SNP using a Trena object
}
\examples{
# Create a Trena object for human, assign a variant, then assess the effects of the variant
trena <- Trena("hg38")

library(MotifDb)
jaspar.human.pfms <- as.list(query(query(MotifDb, "jaspar2016"), "sapiens"))

variant <- "rs3875089" # chr18:26865469  T->C

tbl <- assessSnp(trena, jaspar.human.pfms, variant, shoulder = 3,
pwmMatchMinimumAsPercentage = 65)


}
