% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureTable.R
\name{FeatureTable}
\alias{FeatureTable}
\title{FeatureTable}
\description{
A template for building documented, tested R6 classes
}
\examples{
  rt <- FeatureTable()

import R6
import data.table

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureTable-new}{\code{FeatureTable$new()}}
\item \href{#method-FeatureTable-getTable}{\code{FeatureTable$getTable()}}
\item \href{#method-FeatureTable-setFundamentalRegions}{\code{FeatureTable$setFundamentalRegions()}}
\item \href{#method-FeatureTable-addRegionFeature}{\code{FeatureTable$addRegionFeature()}}
\item \href{#method-FeatureTable-addGeneFeature}{\code{FeatureTable$addGeneFeature()}}
\item \href{#method-FeatureTable-addDistanceToTSS}{\code{FeatureTable$addDistanceToTSS()}}
\item \href{#method-FeatureTable-clone}{\code{FeatureTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this [R6][R6::R6Class] class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$new(target.gene, reference.genome)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.gene}}{character, a recognized gene symbol}

\item{\code{reference.genome}}{character, a recognized genome shorthand code, e.g., "hg38"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new instance o FeatureTable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-getTable"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-getTable}{}}}
\subsection{Method \code{getTable()}}{
accessor for the table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$getTable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-setFundamentalRegions"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-setFundamentalRegions}{}}}
\subsection{Method \code{setFundamentalRegions()}}{
establish the basic genomic regions, typically
  transcription factor binding sites
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$setFundamentalRegions(tbl.regions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl.regions}}{a data.frame with at least chrom, start end columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-addRegionFeature"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-addRegionFeature}{}}}
\subsection{Method \code{addRegionFeature()}}{
annotate each of the intersecting fundamental regions with
  the value of this feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$addRegionFeature(
  tbl.feature,
  feature.genome,
  feature.guide,
  default.values
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl.feature}}{a data.frame with chrom, start and end columns, and values
of interest described in the annotation guidel}

\item{\code{feature.genome}}{character, a recognized genome shorthand code, e.g., "hg38"}

\item{\code{feature.guide}}{a neme list identifying columns of interest in
tbl.bed, and the column names to use for them in the feature table}

\item{\code{default.values}}{list matches feature.guide structure}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-addGeneFeature"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-addGeneFeature}{}}}
\subsection{Method \code{addGeneFeature()}}{
annotate each of TFs, via their associated  motif mapped to
the genome, with the value of the feature described in the feature guide
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$addGeneFeature(tbl.feature, feature.name, default.value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl.feature}}{a data.frame with gene and value columns, values
of interest described in the featureguide}

\item{\code{feature.name}}{character use this string to name the new column}

\item{\code{default.value}}{any the value to assign unmatched genes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-addDistanceToTSS"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-addDistanceToTSS}{}}}
\subsection{Method \code{addDistanceToTSS()}}{
annotate each region with a new column, the distance to TSS
   upstream region starts will be negative, downstream will be positive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$addDistanceToTSS(tss.loc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tss.loc}}{integer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
