% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFilter.R
\docType{class}
\name{FootprintFilter-class}
\alias{FootprintFilter-class}
\alias{.FootprintFilter}
\alias{FootprintFilter}
\title{Create a FootprintFilter object}
\description{
A FootprintFilter object allows for filtering based on gene footprinting databases. Using its
associated \code{getCandidates} method and URIs for both a genome database and project database,
a FootprintFilter object can be used to filter a list of possible transcription factors to those
that match footprint motifs for a supplied target gene.
}
