% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneOntologyFilter.R
\docType{class}
\name{GeneOntologyFilter-class}
\alias{GeneOntologyFilter-class}
\alias{.GeneOntologyFilter}
\alias{GeneOntologyFilter}
\title{Create a GeneOntologyFilter object}
\description{
A GeneOntologyFilter object allows for filtering based on gene ontology (GO) terms. Its
associated \code{getCandidates} method uses an organism database and a GO term (e.g. GO:#######)
to filter a list of possible regulators factors to those that match the GO term.
}
