% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDHSFilter.R
\docType{methods}
\name{geneSymbolToTSS,HumanDHSFilter-method}
\alias{geneSymbolToTSS,HumanDHSFilter-method}
\alias{geneSymbolToTSS}
\title{Get candidate genes using a human DHS filter}
\usage{
\S4method{geneSymbolToTSS}{HumanDHSFilter}(obj, geneSymbol)
}
\arguments{
\item{obj}{An object of class HumanDHSFilter}

\item{geneSymbol}{A gene of interest}
}
\value{
A list, where one element is a character vector of chromosomes in which the geneSymbol
is found and the other is a vector of the transcription start sites of the geneSymbol
}
\description{
Get candidate genes using a human DHS filter
}
\seealso{
\code{\link{HumanDHSFilter}}
}
