% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFilter.R
\docType{methods}
\name{getCandidates,FootprintFilter-method}
\alias{getCandidates,FootprintFilter-method}
\alias{getCandidates-FootprintFilter}
\title{Get candidate genes using the footprint filter}
\usage{
\S4method{getCandidates}{FootprintFilter}(obj)
}
\arguments{
\item{obj}{An object of class FootprintFilter}
}
\value{
A list, where one element is the transcription factors found in the footprints and the
other is a data frame containing all the meta data for the footprints
}
\description{
Get candidate genes using the footprint filter
}
\examples{

# Use footprint filter with the included SQLite database for MEF2C to filter candidates
# in the included Alzheimer's dataset, using the Trena object to get regions
target.gene <- "MEF2C"
db.address <- system.file(package="trena", "extdata")
genome.db.uri <- paste("sqlite:/",db.address,"mef2c.neighborhood.hg38.gtfAnnotation.db", sep = "/")
project.db.uri <- paste("sqlite:/",db.address,"mef2c.neigborhood.hg38.footprints.db", sep = "/")
size.upstream <- 1000
size.downstream <- 1000

# Construct a Trena object and use it to retrive the regions
trena <- Trena("hg38")
regions <- getProximalPromoter(trena,target.gene, size.upstream, size.downstream)

footprint.filter <- FootprintFilter(genomeDB = genome.db.uri, footprintDB = project.db.uri,
regions = regions)
footprints <- getCandidates(footprint.filter)
}
\seealso{
\code{\link{FootprintFilter}}

Other getCandidate Methods: \code{\link{getCandidates,GeneOntologyFilter-method}},
  \code{\link{getCandidates,HumanDHSFilter-method}},
  \code{\link{getCandidates,VarianceFilter-method}},
  \code{\link{getCandidates}}
}
