% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFilter.R
\docType{methods}
\name{getCandidates,FootprintFilter-method}
\alias{getCandidates,FootprintFilter-method}
\alias{getCandidates-FootprintFilter}
\title{Get candidate genes using the variance filter}
\usage{
\S4method{getCandidates}{FootprintFilter}(obj, extraArgs)
}
\arguments{
\item{obj}{An object of class FootprintFilter}

\item{extraArgs}{A named list containing 5 fields:
\itemize{
\item{"target.gene" A designated target gene that should be part of the mtx.assay data}
\item{"genome.db.uri" A connection to a genome database containing footprint information}
\item{"project.db.uri" A connection to a project database containing footprint information}
\item{"size.upstream" An integer denoting the distance upstream of the target gene to look for footprints}
\item{"size.downstream" An integer denoting the distance downstream of the target gene to look for footprints}
}}
}
\value{
A vector containing all genes with variances less than the target gene
}
\description{
Get candidate genes using the variance filter
}
\examples{

# Use footprint filter with the included SQLite database for MEF2C to filter candidates
# in the included Alzheimer's dataset
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
footprint.filter <- FootprintFilter(mtx.assay = mtx.sub)

target.gene <- "MEF2C"
db.address <- system.file(package="TReNA", "extdata")
genome.db.uri <- paste("sqlite:/",db.address,"genome.sub.db", sep = "/")
project.db.uri <- paste("sqlite:/",db.address,"project.sub.db", sep = "/")

tfs <- getCandidates(footprint.filter, extraArgs = list("target.gene" = target.gene,
"genome.db.uri" = genome.db.uri, "project.db.uri" = project.db.uri,
"size.upstream" = 1000, "size.downstream" = 1000))
}
\seealso{
\code{\link{FootprintFilter}}

Other getCandidate Methods: \code{\link{getCandidates,NullFilter-method}},
  \code{\link{getCandidates,VarianceFilter-method}},
  \code{\link{getCandidates}}
}
