% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDHSFilter.R
\docType{methods}
\name{getCandidates,HumanDHSFilter-method}
\alias{getCandidates,HumanDHSFilter-method}
\alias{getCandidates-HumanDHSFilter}
\title{Get candidate genes using a human DHS filter}
\usage{
\S4method{getCandidates}{HumanDHSFilter}(obj)
}
\arguments{
\item{obj}{An object of class FootprintFilter}
}
\value{
A list, where one element a character vector of transcription factors that match
the GO term and the other is an empty data frame.
}
\description{
Get candidate genes using a human DHS filter
}
\examples{

# Make a filter for "transcription, DNA-templated" and use it to filter candidates
\dontrun{
#' load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "VRK2"
promoter.length <- 1000
genomeName <- "hg38"
db.address <- system.file(package="trena", "extdata")
genome.db.uri    <- paste("sqlite:/", db.address, "vrk2.neighborhood.hg38.gtfAnnotation.db",  sep = "/")
jaspar.human <- as.list(query(query(MotifDb, "sapiens"),"jaspar2016"))

# Grab regions for VRK2 using shoulder size of 1000
trena <- Trena(genomeName)
tbl.regions <- getProximalPromoter(trena, "VRK2", 1000, 1000)

hd.filter <- HumanDHSFilter(genomeName, pwmMatchPercentageThreshold = 85,
geneInfoDatabase.uri = genome.db.uri, regions = tbl.regions, pfms = jaspar.human)

getCandidates(hd.filter)
}
}
\seealso{
\code{\link{FootprintFilter}}

Other getCandidate Methods: \code{\link{getCandidates,FootprintFilter-method}},
  \code{\link{getCandidates,GeneOntologyFilter-method}},
  \code{\link{getCandidates,VarianceFilter-method}},
  \code{\link{getCandidates}}
}
\concept{getCandidate Methods}
