% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDHSFilter.R
\docType{methods}
\name{getCandidates,HumanDHSFilter-method}
\alias{getCandidates,HumanDHSFilter-method}
\alias{getCandidates-HumanDHSFilter}
\title{Get candidate genes using a human DHS filter}
\usage{
\S4method{getCandidates}{HumanDHSFilter}(obj)
}
\arguments{
\item{obj}{An object of class FootprintFilter}
}
\value{
A list, where one element a character vector of transcription factors that match
the GO term and the other is an empty data frame.
}
\description{
Get candidate genes using a human DHS filter
}
\examples{

# Make a filter for "transcription, DNA-templated" and use it to filter candidates
#' load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "VRK2"
promoter.length <- 1000
genome <- "hg38"
db.address <- system.file(package="trena", "extdata")
genome.db.uri    <- paste("sqlite:/", db.address, "vrk2.genome.db",  sep = "/")
gene.spec=list(targetGene=target.gene,
tssUpstream=promoter.length,tssDownstream=promoter.length)

hd.filter <- HumanDHSFilter(genome, pwmMatchPercentageThreshold = 85,
geneInfoDatabase.uri = genome.db.uri, geneCenteredSpec = gene.spec)

getCandidates(hd.filter)
}
\seealso{
\code{\link{FootprintFilter}}

Other getCandidate Methods: \code{\link{getCandidates,FootprintFilter-method}},
  \code{\link{getCandidates,GeneOntologyFilter-method}},
  \code{\link{getCandidates,VarianceFilter-method}},
  \code{\link{getCandidates}},
  \code{\link{getRegulatoryRegions,HumanDHSFilter-method}}
}
