% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDHSFilter.R
\docType{methods}
\name{getEncodeRegulatoryTableNames,HumanDHSFilter-method}
\alias{getEncodeRegulatoryTableNames,HumanDHSFilter-method}
\alias{getEncodeRegulatoryTableNames}
\title{Get Encode regulatory tables using a human DHS filter}
\usage{
\S4method{getEncodeRegulatoryTableNames}{HumanDHSFilter}(obj)
}
\arguments{
\item{obj}{An object of class HumanDHSFilter}
}
\value{
A character vector containing the names of the Encode regulatory tables for the regions
contained in the HumanDHSFilter object
}
\description{
Get Encode regulatory tables using a human DHS filter
}
\examples{

load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "VRK2"
promoter.length <- 1000
genome <- "hg38"
db.address <- system.file(package="trena", "extdata")
genome.db.uri    <- paste("sqlite:/", db.address, "vrk2.genome.db",  sep = "/")
gene.spec=list(targetGene=target.gene,
tssUpstream=promoter.length,tssDownstream=promoter.length)

hd.filter <- HumanDHSFilter(genome, pwmMatchPercentageThreshold = 85,
geneInfoDatabase.uri = genome.db.uri, geneCenteredSpec = gene.spec)

getEncodeRegulatoryTableNames(hd.filter)
}
\seealso{
\code{\link{HumanDHSFilter}}
}
