% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDHSFilter.R
\name{getEncodeRegulatoryTableNames}
\alias{getEncodeRegulatoryTableNames}
\title{Get Encode regulatory tables using a human DHS filter}
\usage{
getEncodeRegulatoryTableNames(obj)
}
\arguments{
\item{obj}{An object of class HumanDHSFilter}
}
\value{
A character vector containing the names of the Encode regulatory tables for the regions
contained in the HumanDHSFilter object
}
\description{
Get Encode regulatory tables using a human DHS filter
}
\examples{

load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "VRK2"
promoter.length <- 1000
genome <- "hg38"
db.address <- system.file(package="trena", "extdata")
genome.db.uri    <- paste("sqlite:/", db.address, "vrk2.genome.db",  sep = "/")

# Grab regions for VRK2 using shoulder size of 1000
trena <- Trena(genome)
tbl.regions <- getProximalPromoter(trena, "VRK2", 1000, 1000)

hd.filter <- HumanDHSFilter(genome, pwmMatchPercentageThreshold = 85,
geneInfoDatabase.uri = genome.db.uri, regions = tbl.regions)

getEncodeRegulatoryTableNames(hd.filter) 
}
\seealso{
\code{\link{HumanDHSFilter}}
}
