% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trena.R
\name{getProximalPromoter,Trena-method}
\alias{getProximalPromoter,Trena-method}
\alias{getProximalPromoter}
\title{Grab the region of the proximal promoter for a given gene symbol}
\usage{
\S4method{getProximalPromoter}{Trena}(obj, geneSymbols, tssUpstream = 1000, tssDownstream = 1000)
}
\arguments{
\item{obj}{An object of class Trena}

\item{geneSymbols}{A vector containing genes of interest}

\item{tssUpstream}{A designated distance upstream of the promoter to use as a shoulder
(default = 1000)}

\item{tssDownstream}{A designated distance downstream of the promoter to use as a shoulder
(default = 1000)}
}
\value{
A dataframe containing the regions surrounding the proximal promoter
}
\description{
For the genome of a given Trena object, retrieve a data frame containing the region
surrounding a target gene.
}
\examples{
if(interactive()) {  # too slow for the bioc windows build
   # Retrieve the proximal promoter for MEF2C using a shoulder size of 2000 on each side
   trena <- Trena("hg38")
   regions <- getProximalPromoter(trena, "MEF2C", 2000, 2000)
   }
}
