% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{getRegulators}
\alias{getRegulators}
\title{Retrieve the candiate regulators from a Solver object}
\usage{
getRegulators(obj)
}
\arguments{
\item{obj}{An object of class Solver}
}
\value{
The candidate regulators associated with a Solver object
}
\description{
Retrieve the candiate regulators from a Solver object
}
\examples{
# Create a Solver object using the included Alzheimer's data and retrieve the regulators
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
solver <- Solver(mtx.sub)
mtx <- getRegulators(solver) 
}
