% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDHSFilter.R
\docType{methods}
\name{getRegulatoryRegions,HumanDHSFilter-method}
\alias{getRegulatoryRegions,HumanDHSFilter-method}
\alias{getCandidates-FootprintFilter}
\title{Get candidate genes using a human DHS filter}
\usage{
\S4method{getRegulatoryRegions}{HumanDHSFilter}(obj, encode.table.name,
  chromosome, start, end, score.threshold = 0)
}
\arguments{
\item{obj}{An object of class FootprintFilter}
}
\value{
A list, where one element a character vector of transcription factors that match
the GO term and the other is an empty data frame.
}
\description{
Get candidate genes using a human DHS filter
}
\seealso{
\code{\link{FootprintFilter}}

Other getCandidate Methods: \code{\link{getCandidates,FootprintFilter-method}},
  \code{\link{getCandidates,GeneOntologyFilter-method}},
  \code{\link{getCandidates,HumanDHSFilter-method}},
  \code{\link{getCandidates,VarianceFilter-method}},
  \code{\link{getCandidates}},
  \code{\link{getSequence_tmp,HumanDHSFilter-method}}
}
