% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDHSFilter.R
\name{HumanDHSFilter}
\alias{HumanDHSFilter}
\title{Create a CandidateFilter using Human DNAse Hypersensitivity}
\usage{
HumanDHSFilter(genomeName, encodeTableName = "wgEncodeRegDnaseClustered",
  pwmMatchPercentageThreshold, geneInfoDatabase.uri,
  geneCenteredSpec = list(), regionsSpec = list(),
  variants = NA_character_, quiet = TRUE)
}
\arguments{
\item{genomeName}{A character string indicating the reference genome; currently, the only
accepted strings are "hg38" and "hg19", both of which are human genomes.}

\item{encodeTableName}{(default = "wgEncodeRegDnaseClustered")}

\item{pwmMatchPercentageThreshold}{A numeric from 0-100 to serve as a threshold for a match}

\item{geneInfoDatabase.uri}{An address for a gene database}

\item{geneCenteredSpec}{A gene-centered spec}

\item{regionsSpec}{A regions spec}

\item{variants}{Variants}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A CandidateFilter class object that filters using Human DHS data
}
\description{
Create a CandidateFilter using Human DNAse Hypersensitivity
}
\seealso{
\code{\link{getCandidates-HumanDHSFilter}},

Other Solver class objects: \code{\link{BayesSpikeSolver}},
  \code{\link{EnsembleSolver}},
  \code{\link{LassoPVSolver}}, \code{\link{LassoSolver}},
  \code{\link{PearsonSolver}},
  \code{\link{RandomForestSolver}},
  \code{\link{RidgeSolver}}, \code{\link{Solver-class}},
  \code{\link{SpearmanSolver}},
  \code{\link{SqrtLassoSolver}}
}
