% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSolver.R
\name{RandomForestSolver}
\alias{RandomForestSolver}
\title{Create a Solver class object using the Random Forest solver}
\usage{
RandomForestSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with Random Forest as the solver
}
\description{
Create a Solver class object using the Random Forest solver
}
\examples{
solver <- RandomForestSolver()
}
\seealso{
\code{\link{solve.RandomForest}}, \code{\link{getAssayData}}

Other Solver class objects: \code{\link{BayesSpikeSolver}},
  \code{\link{EnsembleSolver}},
  \code{\link{LassoPVSolver}}, \code{\link{LassoSolver}},
  \code{\link{PearsonSolver}}, \code{\link{RidgeSolver}},
  \code{\link{Solver-class}}, \code{\link{SpearmanSolver}},
  \code{\link{SqrtLassoSolver}}
}
