% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeSolver.R
\name{RidgeSolver}
\alias{RidgeSolver}
\title{Create a Solver class object using the Ridge solver}
\usage{
RidgeSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with Ridge as the solver
}
\description{
Create a Solver class object using the Ridge solver
}
\examples{
solver <- RidgeSolver()
}
\seealso{
\code{\link{solve.Ridge}}, \code{\link{getAssayData}}

Other Solver class objects: \code{\link{BayesSpikeSolver}},
  \code{\link{EnsembleSolver}},
  \code{\link{LassoPVSolver}}, \code{\link{LassoSolver}},
  \code{\link{PearsonSolver}},
  \code{\link{RandomForestSolver}},
  \code{\link{Solver-class}}, \code{\link{SpearmanSolver}},
  \code{\link{SqrtLassoSolver}}
}
