% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDHSFilter.R
\docType{methods}
\name{show,HumanDHSFilter-method}
\alias{show,HumanDHSFilter-method}
\alias{show-HumanDHSFilter}
\title{Show the details of a human DHS filter}
\usage{
\S4method{show}{HumanDHSFilter}(object)
}
\arguments{
\item{object}{An object of class HumanDHSFilter}
}
\value{
A list, where one element a character vector of transcription factors that match
the GO term and the other is an empty data frame.
}
\description{
Show the details of a human DHS filter
}
\seealso{
\code{\link{HumanDHSFilter}}
}
