% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeSolver.R
\docType{methods}
\name{show,RidgeSolver-method}
\alias{show,RidgeSolver-method}
\alias{show.RidgeSolver}
\title{Show the Ridge Solver}
\usage{
\S4method{show}{RidgeSolver}(obj)
}
\arguments{
\item{obj}{An object of the class RidgeSolver}
}
\value{
A truncated view of the supplied object
}
\description{
Show the Ridge Solver
}
\examples{
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
ridge.solver <- RidgeSolver(mtx.sub, target.gene, tfs)
show(ridge.solver)
}
