% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpearmanSolver.R
\docType{methods}
\name{run,SpearmanSolver-method}
\alias{run,SpearmanSolver-method}
\alias{run.SpearmanSolver}
\alias{solve.Spearman}
\title{Run the Spearman Solver}
\usage{
\S4method{run}{SpearmanSolver}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs)), extraArgs = list())
}
\arguments{
\item{obj}{An object of class Solver with "spearman" as the solver string}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}

\item{extraArgs}{Modifiers to the Spearman solver}
}
\value{
The set of Spearman Correlation Coefficients between each transcription factor and the target gene.
}
\description{
Given a TReNA object with Spearman as the solver, use the \code{\link{cor}}
function with \code{method = "spearman"} to esimate coefficients for each transcription factor
as a predictor of the target gene's expression level.
This method should be called using the \code{\link{solve}} method on an appropriate TReNA object.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
trena <- TReNA(mtx.assay = mtx.sub, solver = "pearson")
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
tbl <- solve(trena, target.gene, tfs)
}
\seealso{
\code{\link{cor}}, \code{\link{SpearmanSolver}}

Other solver methods: \code{\link{run,BayesSpikeSolver-method}},
  \code{\link{run,EnsembleSolver-method}},
  \code{\link{run,LassoPVSolver-method}},
  \code{\link{run,LassoSolver-method}},
  \code{\link{run,PearsonSolver-method}},
  \code{\link{run,RandomForestSolver-method}},
  \code{\link{run,RidgeSolver-method}},
  \code{\link{run,SqrtLassoSolver-method}},
  \code{\link{solve,TReNA-method}}
}
