% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\docType{methods}
\name{run,SqrtLassoSolver-method}
\alias{run,SqrtLassoSolver-method}
\alias{run.SqrtLassoSolver}
\alias{solve.SqrtLasso}
\title{Run the Square Root LASSO Solver}
\usage{
\S4method{run}{SqrtLassoSolver}(obj)
}
\arguments{
\item{obj}{An object of class Solver with "sqrtlasso" as the solver string}
}
\value{
A data frame containing the coefficients relating the target gene to
each transcription factor, plus other fit parameters.
}
\description{
Given a TReNA object with Square Root LASSO as the solver,
use the \code{\link{slim}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
This method should be called using the \code{\link{solve}} method on an appropriate TReNA object.
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Square Root LASSO as solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
sqrt.solver <- SqrtLassoSolver(mtx.sub, target.gene, tfs)
tbl <- run(sqrt.solver)

# Solve the same problem but use 8 cores
sqrt.solver <- SqrtLassoSolver(mtx.sub, target.gene, tfs, nCores = 8)
tbl <- run(sqrt.solver)
}
\seealso{
\code{\link{slim}}, \code{\link{SqrtLassoSolver}}

Other solver methods: \code{\link{run,BayesSpikeSolver-method}},
  \code{\link{run,EnsembleSolver-method}},
  \code{\link{run,LassoPVSolver-method}},
  \code{\link{run,LassoSolver-method}},
  \code{\link{run,PearsonSolver-method}},
  \code{\link{run,RandomForestSolver-method}},
  \code{\link{run,RidgeSolver-method}},
  \code{\link{run,SpearmanSolver-method}},
  \code{\link{solve,TReNA-method}}
}
