% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\docType{class}
\name{Solver-class}
\alias{Solver-class}
\alias{.Solver}
\alias{Solver}
\title{Define an object of class Solver}
\usage{
Solver(mtx.assay = matrix(), targetGene, candidateRegulators, quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{targetGene}{A designated target gene that should be part of the mtx.assay data}

\item{candidateRegulators}{The designated set of transcription factors that could be associated}

\item{quiet}{A logical indicating whether or not the Solver object should print output}
}
\value{
An object of the Solver class
}
\description{
The Solver class is a base class that governs the different solvers available in \code{trena}.
It is rarely called by itself; rather, interaction with a particular solver object
is achieved using a specific solver type.
}
\examples{
#' # Create a Solver object using the included Alzheimer's data
load(system.file(package="trena", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
targetGene <- "MEF2C"
candidateRegulators <- setdiff(rownames(mtx.sub), targetGene)
solver <- Solver(mtx.sub, targetGene, candidateRegulators) # Create a simple Solver object with default options

}
\seealso{
\code{\link{getAssayData}}, \code{\link{getTarget}}, \code{\link{getRegulators}}

Other Solver class objects: 
\code{\link{BayesSpikeSolver}},
\code{\link{EnsembleSolver}},
\code{\link{HumanDHSFilter-class}},
\code{\link{LassoPVSolver}},
\code{\link{LassoSolver}},
\code{\link{PearsonSolver}},
\code{\link{RandomForestSolver}},
\code{\link{RidgeSolver}},
\code{\link{SpearmanSolver}},
\code{\link{XGBoostSolver}}
}
\concept{Solver class objects}
