% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\name{SqrtLassoSolver}
\alias{SqrtLassoSolver}
\title{Create a Solver class object using the Square Root LASSO solver}
\usage{
SqrtLassoSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with Square Root LASSO as the solver
}
\description{
Create a Solver class object using the Square Root LASSO solver
}
\examples{
solver <- SqrtLassoSolver()
}
\seealso{
\code{\link{solve.SqrtLasso}}, \code{\link{getAssayData}}

Other Solver class objects: \code{\link{BayesSpikeSolver}},
  \code{\link{EnsembleSolver}},
  \code{\link{LassoPVSolver}}, \code{\link{LassoSolver}},
  \code{\link{PearsonSolver}},
  \code{\link{RandomForestSolver}},
  \code{\link{RidgeSolver}}, \code{\link{Solver-class}},
  \code{\link{SpearmanSolver}}
}
